requirements.txt
opencv-python-headless
face-recognition
cmake
python-multipart
uvicorn
flask
Flask-Cors
psycopg2-binary

